requirements.txt
bleach==6.0.0
bokeh==2.4.3
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
colorcet==3.0.1
distlib==0.3.6
filelock==3.10.0
holoviews==1.15.4
idna==3.4
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
panel==0.14.4
param==1.12.3
Pillow==9.4.0
platformdirs==3.1.1
pyct==0.5.0
python-dateutil==2.8.2
pytz==2022.7.1
pyviz-comms==2.2.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
webencodings==0.5.1
